util: move inspect in separate file
The inspect function became very big and it's better to handle this
in a separate file.

Backport-PR-URL: https://github.com/nodejs/node/pull/23226
PR-URL: https://github.com/nodejs/node/pull/22845
